refactor: エリアコンポーネントからカードコンポーネントに移行。
