+++ components/layouts/LearnSettingAreaComponent.tsx
@@ -1,6 +1,6 @@
-import { Button } from '@chakra-ui/react'
 import React from 'react'
-import { useForm } from 'react-hook-form'
+
+import LearnSettingCardComponent from '../mass/LearnSettingCardComponent'
 
 type Inputs = {
   example: string
@@ -8,29 +8,10 @@ type Inputs = {
 }
 
 const LearnSettingAreaComponent: React.FC = () => {
-  const {
-    register,
-    handleSubmit,
-    watch,
-    formState: { errors },
-  } = useForm<Inputs>()
-
-  const onSubmit = (data: unknown) => console.log(data)
-
-  console.log(watch('example'))
-
   return (
-    <form onSubmit={handleSubmit(onSubmit)}>
-      <button className="btn btn-primary">DaisyUI Button</button>
-      <p className="text-gray-300 bg-gray-100">aaa</p>
-      <div className="p-10">
-        <input defaultValue="test" {...register('example')} />
-        <input {...register('exampleRequired', { required: true })} />
-        {errors.exampleRequired && <span>This field is required</span>}
-      </div>
-      <input type="submit" />
-      <Button colorScheme="blue">Button</Button>
-    </form>
+    <section>
+      <LearnSettingCardComponent />
+    </section>
   )
 }
